==> 1_wqdata/src/4_wqp_lagos_merge.Rmd <==
---
title: "4_wqp_lagos_merge"
author: "Matthew Ross"
date: "6/11/2018"
output:
  html_document:
    toc: true
editor_options: 
  chunk_output_type: console
---


# Merging LAGOS with Water Quality Portal 

In addition to the relativley munging intensive Water Quality Portal data, we are also leveraging a more 'analysis-ready' dataset from the LAke multi-scaled GEoSpatial and temporal dataset (LAGOS). This dataset is a harmonized dataset that includes secchi, doc, and chlorophyll data from datasets in the Water Quality Portal and many other datasets from individual researchers, state agencies, and citizen science groups. 

Merging this dataset with our harmonized, unified, and widened water quality portal data requires some parallel munging but with much less quality control, since the LAGOS crew has already done that. 

```{r setup, include=F, warnings='hide'}
library(feather)
library(tidyverse)
library(knitr)
library(kableExtra)
library(pander)
library(LAGOSNE)
library(lubridate)
library(tidyr)

knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir='../..')


```

```{r}
#Load pluribus function
source('1_wqdata/src/merge_unity_functions.R')
```


## LAGOS data read

The package `LAGOSNE` has all of the lagos data and metadata already read in and requires a few simple commands to download. We will be using version 1.087.1 for this paper

```{r lagos get}
# For the first time you will download the data using lagosne_get
lagosne_get('1.087.1')

#Once that data has been downloaded you can simply read it into R
lagos <- lagosne_load("1.087.1")

nutr <- lagos$epi_nutr


```


### WQP read

We'll read in the wide water quality portal data so we know exactly how to reshape the LAGOS data


```{r}
wqp.wide <- read_feather('1_wqdata/out/wqp_unity_wide.feather')

```



### LAGOS reshaping


#### Secchi 

LAGOS data of interest for this project lives in two key places `lagos$secchi` and `lagos$epi_nutr`. Let's first look at the secchi data, reshape it, and remove any leftover duplicates

```{r}
secchi <- lagos$secchi %>%
  #All lagos data is date only
  mutate(date_only=T) %>%
  #Mutate the date data to a posixct format
  mutate(date_unity=mdy_hms(paste(as.character(sampledate),'00:00:00'),tz='UTC')) %>%
  #Grab only the columns needed
  select(SiteID=lagoslakeid,date_unity,date_only,harmonized_value=secchi)  %>%
  #Add a harmonized paramter column
  mutate(harmonized_parameter='secchi')





#Thats okay, we have a function for that
secchi.unity <- pluribus(secchi,20,'date_unity')

```

#### Chlorophyll and DOC

Chlorophyll and DOC data both live in the `epi_nutr` dataset and need a small amount of duplicate removal

```{r}

#Now we have a complete LAGOS secchi dataset, but LAGOS also holds chlorophyll and DOC data
chl.doc <- lagos$epi_nutr %>%
  mutate(date_unity=mdy_hms(paste(as.character(sampledate),'00:00:00'),tz='UTC')) %>%
  select(SiteID=lagoslakeid,date_unity,chla,doc) %>%
  #lengthen dataset so it plays nicely with unity code
  gather(.,key=harmonized_parameter,value=harmonized_value,-date_unity,-SiteID) %>%
  #Add date_only column
  mutate(date_only=T) %>%
  select(names(secchi))

chl.doc.unity <- pluribus(chl.doc,20,'date_unity')

```

### Lagos unification
```{r}

lagos.unity <- rbind(secchi.unity,chl.doc.unity) %>%
  filter(!is.na(harmonized_value)) %>%
  mutate(SiteID=as.character(SiteID)) %>%
  spread(key=harmonized_parameter,value=harmonized_value) %>%
  rename(chl_a=chla) %>%
  mutate(p_sand=NA,
         tis=NA,
         tss=NA,
         source='LAGOS') %>%
  select(names(wqp.wide))

```



## Water Quality Portal and Lagos unification
```{r}
wqp.lagos <- rbind(wqp.wide,lagos.unity) 


#Setting data thresholds to eliminated non-sensical data. 
# 0.01 > Chl_a < 10^5
# 0.01 > TSS < 10^4
# 0.01 | 0 > secchi < 100
# 0.01 > DOC < 500
wqp.lagos.thresholds <- wqp.lagos %>%
  mutate(chl_a = ifelse(chl_a < 10^5 & chl_a > 0.01,chl_a,NA)) %>%
  mutate(tss = ifelse(tss < 10^6 & tss > 0.01,tss,NA)) %>%
  mutate(secchi = ifelse(secchi < 100 & secchi > 0.01 | secchi == 0,secchi,NA)) %>%
  mutate(doc = ifelse(doc > 0.01 & doc < 500,doc,NA)) %>%
  mutate(p_sand = ifelse(p_sand > 0 & p_sand < 100,p_sand,NA)) %>%
  mutate(tis = ifelse(tis > 0 & tis < 10^5,tis,NA))

write_feather(wqp.lagos.thresholds,path='1_wqdata/out/wqp_lagos_unity.feather')
write_csv(wqp.lagos.thresholds,path='1_wqdata/out/wqp_lagos_unity.csv')


# scipiper::gd_put('1_wqdata/out/wqp_lagos_unity.feather.ind', local_source = '1_wqdata/out/wqp_lagos_unity.feather')
gc()



```
